+++ RPG-Character-Sheets/src/br/com/while42/rpgcs/model/equip/Equipment.java
@@ -3,6 +3,6 @@ package br.com.while42.rpgcs.model.equip;
 import java.io.Serializable;
 
 public interface Equipment extends Serializable {
-	public Integer getCost();
+	public Long getCost();
 	public Double getWeight();
 }
