+++ RPG-Character-Sheets/src/br/com/while42/rpgcs/model/equip/armor/AbstractArmor.java
@@ -16,7 +16,7 @@ public abstract class AbstractArmor implements Armor {
 	}
 
 	@Override
-	public Integer getCost() {
+	public Long getCost() {
 		return armor.getCost();
 	}
